python-sdk/pyproject.toml
[project]
name = "wdsf-python-sdk"
version = "0.1.0"
description = "Python client SDK for World DanceSport Federation API"
authors = [
    {name = "Alibek Cholponbaev", email = "Volume99917@gmail.com"}
]
readme = "README.md"
requires-python = ">= 3.11"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.11.11"
python-dotenv = "^1.1.0"

[tool.ruff]
target-version = "py311"
